+++ apps/desktop/pnpm-workspace.yaml
@@ -1,8 +1,11 @@
 packages:
+  - '../../packages/const'
   - '../../packages/electron-server-ipc'
   - '../../packages/electron-client-ipc'
   - '../../packages/file-loaders'
   - '../../packages/desktop-bridge'
   - '../../packages/device-gateway-client'
   - '../../packages/local-file-shell'
+  - './stubs/business-const'
+  - './stubs/types'
   - '.'
