+++ apps/desktop/src/main/const/store.ts
@@ -1,11 +1,11 @@
 /**
  * Application settings storage related constants
  */
+import { DEFAULT_ELECTRON_DESKTOP_SHORTCUTS } from '@lobechat/const/desktopGlobalShortcuts';
 import type { NetworkProxySettings } from '@lobechat/electron-client-ipc';
 
 import { appStorageDir } from '@/const/dir';
 import { UPDATE_CHANNEL } from '@/modules/updater/configs';
-import { DEFAULT_SHORTCUTS_CONFIG } from '@/shortcuts';
 import type { ElectronMainStore } from '@/types/store';
 
 /**
@@ -35,7 +35,7 @@ export const STORE_DEFAULTS: ElectronMainStore = {
   gatewayUrl: 'https://device-gateway.lobehub.com',
   locale: 'auto',
   networkProxy: defaultProxySettings,
-  shortcuts: DEFAULT_SHORTCUTS_CONFIG,
+  shortcuts: DEFAULT_ELECTRON_DESKTOP_SHORTCUTS,
   storagePath: appStorageDir,
   themeMode: 'system',
   updateChannel: UPDATE_CHANNEL,
