+++ apps/desktop/src/main/controllers/index.ts
@@ -1,6 +1,7 @@
+import type { DesktopHotkeyId } from '@lobechat/const/desktopGlobalShortcuts';
+
 import type { App } from '@/core/App';
 import { IoCContainer } from '@/core/infrastructure/IoCContainer';
-import type { ShortcutActionType } from '@/shortcuts';
 import { IpcService } from '@/utils/ipc';
 
 const shortcutDecorator = (name: string) => (target: any, methodName: string, descriptor?: any) => {
@@ -15,7 +16,7 @@ const shortcutDecorator = (name: string) => (target: any, methodName: string, de
 /**
  *  shortcut inject decorator
  */
-export const shortcut = (method: ShortcutActionType) => shortcutDecorator(method);
+export const shortcut = (method: DesktopHotkeyId) => shortcutDecorator(method);
 
 const protocolDecorator =
   (urlType: string, action: string) => (target: any, methodName: string, descriptor?: any) => {
