+++ apps/desktop/src/main/core/ui/ShortcutManager.ts
@@ -1,6 +1,6 @@
+import { DEFAULT_ELECTRON_DESKTOP_SHORTCUTS } from '@lobechat/const/desktopGlobalShortcuts';
 import { globalShortcut } from 'electron';
 
-import { DEFAULT_SHORTCUTS_CONFIG } from '@/shortcuts';
 import { createLogger } from '@/utils/logger';
 
 import type { App } from '../App';
@@ -77,8 +77,8 @@ export class ShortcutManager {
     try {
       logger.debug(`Updating shortcut ${id} to ${accelerator}`);
 
-      // 1. Check if ID is valid
-      if (!DEFAULT_SHORTCUTS_CONFIG[id]) {
+      // 1. Check if ID is valid (value may be empty string when disabled by default)
+      if (!(id in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS)) {
         logger.error(`Invalid shortcut ID: ${id}`);
         return { errorType: 'INVALID_ID', success: false };
       }
@@ -231,15 +231,15 @@ export class ShortcutManager {
       // If no configuration, use default configuration
       if (!config || Object.keys(config).length === 0) {
         logger.debug('No shortcuts config found, using defaults');
-        this.shortcutsConfig = { ...DEFAULT_SHORTCUTS_CONFIG };
+        this.shortcutsConfig = { ...DEFAULT_ELECTRON_DESKTOP_SHORTCUTS };
         this.saveShortcutsConfig();
       } else {
-        // Filter out invalid shortcuts that are not in DEFAULT_SHORTCUTS_CONFIG
+        // Filter out invalid shortcuts that are not in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS
         const filteredConfig: Record<string, string> = {};
         let hasInvalidKeys = false;
 
         Object.entries(config).forEach(([id, accelerator]) => {
-          if (DEFAULT_SHORTCUTS_CONFIG[id]) {
+          if (id in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS) {
             filteredConfig[id] = accelerator;
           } else {
             hasInvalidKeys = true;
@@ -248,7 +248,7 @@ export class ShortcutManager {
         });
 
         // Ensure all default shortcuts are present
-        Object.entries(DEFAULT_SHORTCUTS_CONFIG).forEach(([id, defaultAccelerator]) => {
+        Object.entries(DEFAULT_ELECTRON_DESKTOP_SHORTCUTS).forEach(([id, defaultAccelerator]) => {
           if (!(id in filteredConfig)) {
             filteredConfig[id] = defaultAccelerator;
             logger.debug(`Adding missing default shortcut: ${id} = ${defaultAccelerator}`);
@@ -267,7 +267,7 @@ export class ShortcutManager {
       logger.debug('Loaded shortcuts config:', this.shortcutsConfig);
     } catch (error) {
       logger.error('Error loading shortcuts config:', error);
-      this.shortcutsConfig = { ...DEFAULT_SHORTCUTS_CONFIG };
+      this.shortcutsConfig = { ...DEFAULT_ELECTRON_DESKTOP_SHORTCUTS };
       this.saveShortcutsConfig();
     }
   }
@@ -295,9 +295,9 @@ export class ShortcutManager {
     Object.entries(this.shortcutsConfig).forEach(([id, accelerator]) => {
       logger.debug(`Registering shortcut '${id}' with ${accelerator}`);
 
-      // Only register shortcuts that exist in DEFAULT_SHORTCUTS_CONFIG
-      if (!DEFAULT_SHORTCUTS_CONFIG[id]) {
-        logger.debug(`Skipping shortcut '${id}' - not found in DEFAULT_SHORTCUTS_CONFIG`);
+      // Only register shortcuts that exist in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS
+      if (!(id in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS)) {
+        logger.debug(`Skipping shortcut '${id}' - not found in DEFAULT_ELECTRON_DESKTOP_SHORTCUTS`);
         return;
       }
