--- apps/desktop/src/main/shortcuts/index.ts
@@ -1,1 +1,0 @@
-export * from './config';
