+++ apps/desktop/stubs/business-const/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@lobechat/business-const",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "main": "./src/index.ts"
+}
