+++ apps/desktop/stubs/business-const/src/index.ts
@@ -1,0 +1,6 @@
+export const BRANDING_LOGO_URL = '';
+export const BRANDING_NAME = 'LobeHub';
+export const DEFAULT_EMBEDDING_PROVIDER = 'openai';
+export const DEFAULT_MINI_PROVIDER = 'openai';
+export const DEFAULT_PROVIDER = 'openai';
+export const ORG_NAME = 'LobeHub';
