+++ apps/desktop/stubs/types/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@lobechat/types",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "main": "./src/index.ts"
+}
