+++ apps/desktop/stubs/types/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Desktop isolated workspace stub.
+ *
+ * `@lobechat/types` is only consumed via `import type` in desktop code; those
+ * specifiers are erased at build time, so this package needs no runtime
+ * exports.
+ */
+export {};
