+++ packages/const/package.json
@@ -4,7 +4,9 @@
   "private": true,
   "exports": {
     ".": "./src/index.ts",
-    "./currency": "./src/currency.ts"
+    "./currency": "./src/currency.ts",
+    "./desktopGlobalShortcuts": "./src/desktopGlobalShortcuts.ts",
+    "./hotkeys": "./src/hotkeys.ts"
   },
   "main": "./src/index.ts",
   "dependencies": {
