+++ packages/const/src/desktopGlobalShortcuts.ts
@@ -1,0 +1,59 @@
+const combineKeys = (keys: string[]) => keys.join('+');
+
+export const DesktopHotkeyEnum = {
+  OpenSettings: 'openSettings',
+  ShowApp: 'showApp',
+} as const;
+
+export type DesktopHotkeyId = (typeof DesktopHotkeyEnum)[keyof typeof DesktopHotkeyEnum];
+
+export interface DesktopHotkeyItem {
+  id: DesktopHotkeyId;
+  keys: string;
+  nonEditable?: boolean;
+}
+
+export type DesktopHotkeyConfig = Record<DesktopHotkeyId, string>;
+
+interface DesktopGlobalShortcutDefault {
+  /** Electron `globalShortcut` accelerator; empty string means unregistered. */
+  electronAccelerator: string;
+  id: DesktopHotkeyId;
+  nonEditable?: boolean;
+  /** React-hotkey style string for renderer (HotkeyInput, merge defaults). */
+  uiKeys: string;
+}
+
+/**
+ * Single source of truth for desktop (Electron) global shortcut defaults.
+ * Main process reads `electronAccelerator`; renderer uses `uiKeys` and metadata.
+ */
+export const DESKTOP_GLOBAL_SHORTCUT_DEFAULTS = [
+  {
+    electronAccelerator: '',
+    id: DesktopHotkeyEnum.ShowApp,
+    uiKeys: '',
+  },
+  {
+    electronAccelerator: 'CommandOrControl+,',
+    id: DesktopHotkeyEnum.OpenSettings,
+    nonEditable: true,
+    uiKeys: combineKeys(['mod', 'comma']),
+  },
+] as const satisfies readonly DesktopGlobalShortcutDefault[];
+
+export const DESKTOP_HOTKEYS_REGISTRATION: DesktopHotkeyItem[] =
+  DESKTOP_GLOBAL_SHORTCUT_DEFAULTS.map((item): DesktopHotkeyItem => {
+    const base: DesktopHotkeyItem = {
+      id: item.id,
+      keys: item.uiKeys,
+    };
+
+    return 'nonEditable' in item && item.nonEditable ? { ...base, nonEditable: true } : base;
+  });
+
+export const DEFAULT_ELECTRON_DESKTOP_SHORTCUTS: DesktopHotkeyConfig =
+  DESKTOP_GLOBAL_SHORTCUT_DEFAULTS.reduce<DesktopHotkeyConfig>((acc, item) => {
+    acc[item.id] = item.electronAccelerator;
+    return acc;
+  }, {} as DesktopHotkeyConfig);
