+++ packages/const/src/settings/hotkey.ts
@@ -1,6 +1,6 @@
-import type { UserHotkeyConfig } from '@lobechat/types';
+import { type HotkeyId, HOTKEYS_REGISTRATION } from '../hotkeys';
 
-import { HOTKEYS_REGISTRATION } from '../hotkeys';
+type UserHotkeyConfig = Record<HotkeyId, string>;
 
 export const DEFAULT_HOTKEY_CONFIG: UserHotkeyConfig = HOTKEYS_REGISTRATION.reduce(
   (acc: UserHotkeyConfig, item) => {
