+++ src/features/ChatInput/InputEditor/Placeholder.tsx
@@ -1,5 +1,5 @@
-import { KeyEnum } from '@lobechat/types';
-import { combineKeys,Flexbox, Hotkey } from '@lobehub/ui';
+import { KeyEnum } from '@lobechat/const/hotkeys';
+import { combineKeys, Flexbox, Hotkey } from '@lobehub/ui';
 import { memo } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
