+++ src/features/HotkeyHelperPanel/index.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { HotkeyGroupEnum } from '@lobechat/const/hotkeys';
 import { Grid, Icon, Modal, Segmented } from '@lobehub/ui';
 import { MessageSquare, Settings2 } from 'lucide-react';
 import { memo, useState } from 'react';
@@ -7,7 +8,6 @@ import { useTranslation } from 'react-i18next';
 
 import { useGlobalStore } from '@/store/global';
 import { type HotkeyGroupId } from '@/types/hotkey';
-import { HotkeyGroupEnum } from '@/types/hotkey';
 
 import HotkeyContent from './HotkeyContent';
