+++ src/hooks/useHotkeys/globalScope.ts
@@ -1,5 +1,5 @@
 import { INBOX_SESSION_ID } from '@lobechat/const';
-import { HotkeyEnum } from '@lobechat/types';
+import { HotkeyEnum } from '@lobechat/const/hotkeys';
 import { useLocation } from 'react-router-dom';
 
 import { useNavigateToAgent } from '@/hooks/useNavigateToAgent';
